ff2asds

+++ src/main/java/edu/upb/webpool/web/rest/PoolEntryResource.java
@@ -1,21 +1,29 @@
 package edu.upb.webpool.web.rest;
 
+import com.netflix.discovery.converters.Auto;
 import edu.upb.webpool.domain.Pool;
 import edu.upb.webpool.domain.PoolEntry;
 import edu.upb.webpool.repository.PoolEntryRepository;
+import edu.upb.webpool.repository.PoolRepository;
 import edu.upb.webpool.web.rest.errors.BadRequestAlertException;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
+
+import edu.upb.webpool.web.rest.utils.OtpValidator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import tech.jhipster.web.util.HeaderUtil;
 import tech.jhipster.web.util.ResponseUtil;
+
+import javax.servlet.http.HttpServletRequest;
 
 /**
  * REST controller for managing {@link edu.upb.webpool.domain.PoolEntry}.
@@ -33,6 +41,9 @@ public class PoolEntryResource {
 
     private final PoolEntryRepository poolEntryRepository;
 
+    @Autowired
+    private PoolRepository poolRepository;
+
     public PoolEntryResource(PoolEntryRepository poolEntryRepository) {
         this.poolEntryRepository = poolEntryRepository;
     }
@@ -45,11 +56,24 @@ public class PoolEntryResource {
      * @throws URISyntaxException if the Location URI syntax is incorrect.
      */
     @PostMapping("/pool-entries")
-    public ResponseEntity<PoolEntry> createPoolEntry(@RequestBody PoolEntry poolEntry) throws URISyntaxException {
+    public ResponseEntity<PoolEntry> createPoolEntry(@RequestBody PoolEntry poolEntry, HttpServletRequest request) throws Exception {
         log.debug("REST request to save PoolEntry : {}", poolEntry);
         if (poolEntry.getId() != null) {
             throw new BadRequestAlertException("A new poolEntry cannot already have an ID", ENTITY_NAME, "idexists");
         }
+
+        Pool pool = poolRepository.findById(poolEntry.getPool()).orElse(null);
+
+        if(pool == null) {
+            throw new Exception();
+        }
+
+        if(pool.isOtp()) {
+            if(!OtpValidator.validate(poolEntry.getOtp(), SecurityContextHolder.getContext().getAuthentication().getName())) {
+                throw new Exception();
+            }
+        }
+
         poolEntryRepository.findByPoolAndOwner(poolEntry.getPool(), poolEntry.getOwner()).stream().forEach(p -> poolEntryRepository.delete(p));
         PoolEntry result = poolEntryRepository.save(poolEntry);
         return ResponseEntity
